whatsapp_service/requirements.txt
aiomysql==0.2.0
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
asyncpg==0.30.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
cryptography==44.0.0
databases==0.9.0
dnspython==2.7.0
email_validator==2.2.0
Faker==33.1.0
fastapi==0.115.5
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
locust==2.32.5
MarkupSafe==3.0.2
msgpack==1.1.0
prometheus_client==0.21.1
psutil==6.1.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.7.0
pydantic_core==2.27.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pyzmq==26.2.0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
